Utilized Workspace helper for querying for workspace id, code is centralized, turned async, and updated in a few spots
