+++ CSharp/DevVmPowershell/Helpers/IImportApiHelper.cs
@@ -5,6 +5,5 @@ namespace Helpers
 	public interface IImportApiHelper
 	{
 		Task<int> AddDocumentsToWorkspace(int workspaceId, string fileType, int count, string resourceFolderPath);
-		Task<int> GetFirstWorkspaceIdQueryAsync(string workspaceName);
 	}
 }
